testutil/asserter: move asserter and exclude funcs to their own packages
